GIT-1630: refactor singleton helper
