+++ sanic/mixins/signals.py
@@ -1,18 +1,8 @@
 import typing as t
 
 from sanic.exceptions import InvalidSignalFormatException
-from sanic.signals import SignalRegistry, SignalContext, SignalData
-
-
-class Singleton(type):
-    _instances = {}
-
-    def __call__(cls, *args, **kwargs):
-        if cls not in cls._instances:
-            cls._instances[cls] = super(Singleton, cls).__call__(
-                *args, **kwargs
-            )
-        return cls._instances[cls]
+from sanic.signals import SignalRegistry, SignalContext
+from sanic.utils import Singleton
 
 
 class SignalMixin:
